back-end/requirements.txt
wheel
django
django-tastypie
psycopg2-binary
django-cors-headers
django-cors-middleware
